Remove iOS-specific resources and logic
- Deleted iOS project files, including assets, configurations, and entry points.
- Removed `run_ios.sh` script and iOS-related build configurations.
- Cleared out `iosApp` directory and associated dependencies from the project.
