--- demo/iosApp/Configuration/Config.xcconfig
@@ -1,3 +1,0 @@
-TEAM_ID=
-BUNDLE_ID=io.github.kdroidfilter.knotify.demo
-APP_NAME=KnotifyDemo
