--- demo/iosApp/iosApp/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,14 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "app-icon-1024.png",
-      "idiom" : "universal",
-      "platform" : "ios",
-      "size" : "1024x1024"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
