--- demo/iosApp/iosApp/Info.plist
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-    <key>CADisableMinimumFrameDurationOnPhone</key>
-    <true/>
-    <key>NSAppTransportSecurity</key>
-    <dict>
-        <key>NSAllowsArbitraryLoads</key>
-        <true/>
-    </dict>
-    <key>NSFileAccessUsageDescription</key>
-    <string>This app needs access to files to play local videos</string>
-    <key>UIFileSharingEnabled</key>
-    <true/>
-    <key>NSRemoteNotificationUsageDescription</key>
-    <string>This app needs to send notifications to alert you about important events</string>
-    <key>UIBackgroundModes</key>
-    <array>
-        <string>remote-notification</string>
-    </array>
-</dict>
-</plist>
