--- demo/iosApp/iosApp/iosApp.swift
@@ -1,19 +1,0 @@
-import UIKit
-import ComposeApp
-
-@main
-class AppDelegate: UIResponder, UIApplicationDelegate {
-    var window: UIWindow?
-
-    func application(
-        _ application: UIApplication,
-        didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
-    ) -> Bool {
-        window = UIWindow(frame: UIScreen.main.bounds)
-        if let window = window {
-            window.rootViewController = MainKt.MainViewController()
-            window.makeKeyAndVisible()
-        }
-        return true
-    }
-}
